pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "faster-whisper>=1.2.1",
    "kokoro-onnx>=0.5.0",
    "livekit-agents>=1.4.5",
    "livekit-plugins-openai>=1.4.5",
    "livekit-plugins-silero>=1.4.5",
    "llama-cpp-python>=0.3.16",
    "llama-index-llms-llama-cpp>=0.6.0",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "soundfile>=0.13.1",
    "transformers>=5.3.0",
    "uvicorn>=0.41.0",
]
